# api-spec.json
{"openapi":"3.0.0","paths":{"/v1/resumes":{"get":{"operationId":"get","parameters":[],"responses":{"200":{"description":""}},"tags":["resumes"],"security":[{"bearer":[]}]},"post":{"operationId":"createDraft","parameters":[],"requestBody":{"required":true,"description":"Create a resume draft.","content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/ResumeCreateDraftDto"}}}},"responses":{"201":{"description":""}},"tags":["resumes"],"security":[{"bearer":[]}]}},"/v1/resumes/{id}":{"get":{"operationId":"getResumeById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["resumes"],"security":[{"bearer":[]}]}},"/v1/resumes/{id}/version":{"get":{"operationId":"getVersions","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["resumes"],"security":[{"bearer":[]}]}},"/v1/resumes/{id}/version/{version}":{"patch":{"operationId":"updateVersion","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"version","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResumeVersionUpdateDto"}}}},"responses":{"200":{"description":""}},"tags":["resumes"],"security":[{"bearer":[]}]},"get":{"operationId":"getVersion","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"version","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["resumes"],"security":[{"bearer":[]}]}},"/v1/billing":{"get":{"operationId":"get","parameters":[],"responses":{"200":{"description":""}},"tags":["billing"]}},"/v1/users/{id}":{"get":{"operationId":"getUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["users"],"security":[{"bearer":[]}]}}},"info":{"title":"vResume API","description":"Client Guest Credentials [guest@vresume.dev, Cookies123$b]","version":"0.1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"ResumeVersionUpdateDto":{"type":"object","properties":{"query":{"type":"string","description":"The query message of what to update"},"selected":{"type":"string","description":"The selected ui nodes to update"}},"required":["query"]},"ResumeCreateDraftDto":{"type":"object","properties":{"posting":{"type":"string","description":"The job posting to tailor the resume to"},"resume":{"type":"string","format":"binary","description":"The original resume file to use as a knowledge base"}},"required":["posting","resume"]}}},"security":[{"bearer":[]}]}